[960] debug unit tests 4

+++ tofu/tests/tests01_geom/test_01_GG.py
@@ -881,7 +881,8 @@ def test13_LOS_PInOut():
     assert np.allclose(IOut[2, :], Iout)
     npts_vp = VP.shape[1]
     out = GG.LOS_Calc_kMinkMax_VesStruct(Ds, us,
-                                         [VP, VP, VP], [VIn, VIn, VIn], 3,
+                                         [VP, VP, VP], [VIn, VIn, VIn],
+                                         np.int64(3),
                                          np.r_[npts_vp, npts_vp, npts_vp])
     kmin_res = out[0]
     kmax_res = out[1]
